A lot of code cleaning up, LeadAI refactoring

+++ src_rebuild/GAME/C/CONVERT.C
@@ -3,15 +3,6 @@
 #include "DRAW.H"
 #include "CAMERA.H"
 
-// decompiled code
-// original method signature: 
-// void /*$ra*/ Calc_Object_MatrixYZX(MATRIX *mat /*$s0*/, SVECTOR *angles /*$s1*/)
- // line 149, offset 0x0002d3f8
-	/* begin block 1 */
-		// Start line: 298
-	/* end block 1 */
-	// End Line: 299
-
 // [D] [T]
 void Calc_Object_MatrixYZX(MATRIX *mat, SVECTOR *angles)
 {
@@ -32,33 +23,11 @@ void Calc_Object_MatrixYZX(MATRIX *mat, SVECTOR *angles)
 	RotMatrixY(angles->vy, mat);
 }
 
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ _RotMatrixX(MATRIX *m /*$v0*/, short ang /*$a1*/)
- // line 182, offset 0x0002d470
-	/* begin block 1 */
-		// Start line: 508
-	/* end block 1 */
-	// End Line: 509
-
 // [D] [T]
 void _RotMatrixX(MATRIX *m, short ang)
 {
 	RotMatrixX(ang, m);
 }
-
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ _RotMatrixY(MATRIX *m /*$v0*/, short ang /*$a1*/)
- // line 187, offset 0x0002d49c
-	/* begin block 1 */
-		// Start line: 519
-	/* end block 1 */
-	// End Line: 520
 
 // [D] [T]
 void _RotMatrixY(MATRIX *m, short ang)
@@ -66,59 +35,17 @@ void _RotMatrixY(MATRIX *m, short ang)
 	RotMatrixY(ang, m);
 }
 
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ _RotMatrixZ(MATRIX *m /*$v0*/, short ang /*$a1*/)
- // line 192, offset 0x0002d4c8
-	/* begin block 1 */
-		// Start line: 530
-	/* end block 1 */
-	// End Line: 531
-
 // [D] [T]
 void _RotMatrixZ(MATRIX *m, short ang)
 {
 	RotMatrixZ(ang, m);
 }
 
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ RotMatrixXYZ(MATRIX *m /*$v0*/, SVECTOR *r /*$a1*/)
- // line 199, offset 0x0002d4f4
-	/* begin block 1 */
-		// Start line: 545
-	/* end block 1 */
-	// End Line: 546
-
 // [D] [T]
 void RotMatrixXYZ(MATRIX *m, SVECTOR *r)
 {
 	RotMatrix(r, m);
 }
-
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ _MatrixRotate(VECTOR *pos /*$s0*/)
- // line 205, offset 0x0002d51c
-	/* begin block 1 */
-		// Start line: 206
-		// Start offset: 0x0002D51C
-		// Variables:
-	// 		VECTOR temp; // stack offset -24
-	/* end block 1 */
-	// End offset: 0x0002D55C
-	// End Line: 214
-
-	/* begin block 2 */
-		// Start line: 558
-	/* end block 2 */
-	// End Line: 559
 
 // [D] [T]
 void _MatrixRotate(VECTOR *pos)
@@ -131,27 +58,6 @@ void _MatrixRotate(VECTOR *pos)
 	pos->vy = temp.vy;
 	pos->vz = temp.vz;
 }
-
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ InvertMatrix(MATRIX *a /*$a0*/, MATRIX *b /*$a1*/)
- // line 240, offset 0x0002d55c
-	/* begin block 1 */
-		// Start line: 620
-	/* end block 1 */
-	// End Line: 621
-
-	/* begin block 2 */
-		// Start line: 630
-	/* end block 2 */
-	// End Line: 631
-
-	/* begin block 3 */
-		// Start line: 632
-	/* end block 3 */
-	// End Line: 633
 
 // [D] [T]
 void InvertMatrix(MATRIX *a, MATRIX *b)
@@ -166,38 +72,6 @@ void InvertMatrix(MATRIX *a, MATRIX *b)
 	b->m[2][1] = a->m[1][2];
 	b->m[2][2] = a->m[2][2];
 }
-
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ BuildWorldMatrix()
- // line 272, offset 0x0002d304
-	/* begin block 1 */
-		// Start line: 274
-		// Start offset: 0x0002D304
-		// Variables:
-	// 		MATRIX newmatrix; // stack offset -48
-	/* end block 1 */
-	// End offset: 0x0002D3F8
-	// End Line: 289
-
-	/* begin block 2 */
-		// Start line: 544
-	/* end block 2 */
-	// End Line: 545
-
-	/* begin block 3 */
-		// Start line: 545
-	/* end block 3 */
-	// End Line: 546
-
-	/* begin block 4 */
-		// Start line: 547
-	/* end block 4 */
-	// End Line: 548
-
-/* WARNING: Unknown calling convention yet parameter storage is locked */
 
 // [D] [T]
 void BuildWorldMatrix(void)
@@ -231,43 +105,6 @@ void BuildWorldMatrix(void)
 	RotMatrixY(-(camera_angle.vy & 0xfff), &face_camera_work);
 	MulMatrix0(&inv_camera_matrix, &face_camera_work, &face_camera);
 }
-
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ ScaleCamera()
- // line 334, offset 0x0002d5c8
-	/* begin block 1 */
-		// Start line: 336
-		// Start offset: 0x0002D5C8
-		// Variables:
-	// 		MATRIX temp; // stack offset -160
-	// 		MATRIX temp2; // stack offset -128
-	// 		MATRIX scale; // stack offset -96
-	// 		MATRIX scaledcammat; // stack offset -64
-	// 		VECTOR pos; // stack offset -32
-	// 		SVECTOR tempang; // stack offset -16
-	/* end block 1 */
-	// End offset: 0x0002D678
-	// End Line: 361
-
-	/* begin block 2 */
-		// Start line: 818
-	/* end block 2 */
-	// End Line: 819
-
-	/* begin block 3 */
-		// Start line: 819
-	/* end block 3 */
-	// End Line: 820
-
-	/* begin block 4 */
-		// Start line: 831
-	/* end block 4 */
-	// End Line: 832
-
-/* WARNING: Unknown calling convention yet parameter storage is locked */
 
 // [D] [T]
 void ScaleCamera(void)
@@ -306,27 +143,6 @@ void ScaleCamera(void)
 	SetTransMatrix(&scaledcammat);
 }
 
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ Getlong(char *dest /*$a0*/, char *source /*$a1*/)
- // line 365, offset 0x0002d678
-	/* begin block 1 */
-		// Start line: 899
-	/* end block 1 */
-	// End Line: 900
-
-	/* begin block 2 */
-		// Start line: 904
-	/* end block 2 */
-	// End Line: 905
-
-	/* begin block 3 */
-		// Start line: 906
-	/* end block 3 */
-	// End Line: 907
-
 // [D] [T]
 void Getlong(char *dest, char *source)
 {
@@ -336,47 +152,17 @@ void Getlong(char *dest, char *source)
 	dest[3] = source[3];
 }
 
-
-
-// decompiled code
-// original method signature: 
-// void /*$ra*/ RandomInit(long i1 /*$a0*/, long i2 /*$a1*/)
- // line 387, offset 0x0002d6a8
-	/* begin block 1 */
-		// Start line: 389
-		// Start offset: 0x0002D6A8
-		// Variables:
-	// 		int count; // $s0
-	/* end block 1 */
-	// End offset: 0x0002D724
-	// End Line: 402
-
-	/* begin block 2 */
-		// Start line: 948
-	/* end block 2 */
-	// End Line: 949
-
-	/* begin block 3 */
-		// Start line: 949
-	/* end block 3 */
-	// End Line: 950
-
-	/* begin block 4 */
-		// Start line: 952
-	/* end block 4 */
-	// End Line: 953
-
-static long rseed[17]; // offset 0xBD510
+static int rseed[17]; // offset 0xBD510
 static int randomindex = 0;
 static int randomcounter = 0;
 
 // [D] [T]
-void RandomInit(long i1, long i2)
+void RandomInit(int i1, int i2)
 {
 	int step;
 	int count;
 
-	long *sd;
+	int*sd;
 
 	step = 0x3b1cb49;
 	
@@ -399,40 +185,10 @@ void RandomInit(long i1, long i2)
 		Random2(step++);
 }
 
-
-
-// decompiled code
-// original method signature: 
-// long /*$ra*/ Random2(int step /*$a0*/)
- // line 414, offset 0x0002d724
-	/* begin block 1 */
-		// Start line: 416
-		// Start offset: 0x0002D724
-		// Variables:
-	// 		unsigned int seed; // $a0
-	/* end block 1 */
-	// End offset: 0x0002D78C
-	// End Line: 420
-
-	/* begin block 2 */
-		// Start line: 999
-	/* end block 2 */
-	// End Line: 1000
-
-	/* begin block 3 */
-		// Start line: 1012
-	/* end block 3 */
-	// End Line: 1013
-
-	/* begin block 4 */
-		// Start line: 1013
-	/* end block 4 */
-	// End Line: 1014
-
 extern int frameStart;
 
 // [D] [T]
-long Random2(int step)
+int Random2(int step)
 {
 	return (CameraCnt - frameStart) * (CameraCnt - frameStart) * 0x19660d + 0x3c6ef35fU >> 8 & 0xffff;
 }
